requirements.txt
behave
pyyaml
pytest==7.2.1
flatten_json
deepdiff
behavex
behave-html-formatter
opencv-python
numpy
SimpleITK